source-alpaca(refactor): rename worker and separate driver & pull code
